requirements_local.txt
asgiref==3.2.10
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
djangorestframework==3.11.0
entrypoints==0.3
gunicorn==20.0.4
importlib-metadata==1.7.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psycopg2==2.8.5
Pygments==2.6.1
pylint==2.4.4
pyrsistent==0.16.0
python-dateutil==2.8.1
QtPy==1.9.0
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0


# Extra Extensions Used
django-allauth
django-crispy-forms
django-webpack-loader
django-filter
django-phonenumber-field
django-celery-beat
django-admin-honeypot
django-extensions
django-rest-auth
django-secure
fitbit
Faker
phonenumbers